requirements.txt
django==2.2.10
djangorestframework==3.9.2
requests==2.23.0
python==3.7.6